improve code quality and fix images < bug organisation
