+++ src/main/java/game/SurvivalDriver.java
@@ -32,21 +32,11 @@ public class SurvivalDriver extends Driver {
     private static GameScreen gameScreen;
     private static String name;
     private static Leaderboard lb = new Leaderboard();
-    
-    private static int allBubbles = 1;
-    private static int bubbleNumber = 1;
-    private static int spawnTime = 7000;
-    private static Bubble bubble1;
-    private static Bubble bubble2;
-    private static Bubble bubble3;
-    private static Bubble bubble4;
-    private static Bubble bubble5;
 
-    public SurvivalDriver(String name)
-    {
+    public SurvivalDriver(String name) {
         this.name = name;
     }
-    
+
     /**
      * Frame to start the game.
      */
@@ -55,71 +45,15 @@ public class SurvivalDriver extends Driver {
         gameScreen.startGame();
         game.gameStart();
     }
-    
-    /**
-     * Give random generated x and y-coordinates for the bubbles.
-     * @param bubbleNumber The bubble sort which will be generated.
-     */
-    public static void randomPlacedBubble(int bubbleNumber) {
-        int randX = MathFunctions.randomInt(10, Settings.getLevelWidth() - 10);
-        int randY = MathFunctions.randomInt(10, Settings.getLevelHeight() - 200);
-        
-        switch (bubbleNumber) {
-        case 1:
-            bubble1 = new Bubblex8(randX, randY, false, false);
-            break;
-        case 2:
-            bubble2 = new Bubblex16(randX, randY, false, false);
-            break;
-        case 3:
-            bubble3 = new Bubblex32(randX, randY, false, false);
-            break;
-        case 4:
-            bubble4 = new Bubblex64(randX, randY, false, false);
-            break;
-        case 5:
-            bubble5 = new Bubblex128(randX, randY, false, false);
-            break;
-        default:
-            bubble1 = new Bubblex8(randX, randY, false, false);
-            break;
-        }
-    }
-
-    /**
-     * Gives the starting frame.
-     */
-    public void startScreen() {
-        try {
-            new StartScreen(driver);
-        } catch (UnsupportedAudioFileException e1) {
-            e1.printStackTrace();
-        } catch (IOException e1) {
-            e1.printStackTrace();
-        } catch (LineUnavailableException e1) {
-            e1.printStackTrace();
-        }
-        Logger.log("Start screen created", 9, 4);
-    }
 
     /**
      * Check if the game has been lost.
+     * 
      * @return Gives true if the game has ended, otherwise gives false.
      */
     public static boolean checkGameLost() {
         int livesLeft = game.getLives();
         if (livesLeft == 0 && game.inProgress()) {
-            allBubbles = 1;
-            bubbleNumber = 1;
-            spawnTime = 7000;
-            endScore es = new endScore(name , Score.getScore());
-            lb.addScore(es);
-            lb.appendToFile();
-            Score.resetScore();
-            gameScreen.dispose();
-            game.toggleProgress();
-            new LeaderBoardScreen(lb);
-            new LosingScreen(driver, es);
             Logger.log("Game lost", 7, 4);
             return true;
         }
@@ -127,7 +61,25 @@ public class SurvivalDriver extends Driver {
     }
 
     /**
+     * Check if the game has been lost.
+     * 
+     * @return Gives true if the game has ended, otherwise gives false.
+     */
+    public static void gameLost() {
+        game.gameLost();
+        endScore es = new endScore(name, Score.getScore());
+        lb.addScore(es);
+        lb.appendToFile();
+        Score.resetScore();
+        gameScreen.dispose();
+        game.toggleProgress();
+        new LeaderBoardScreen(lb);
+        new LosingScreen(driver, es);
+    }
+
+    /**
      * Setup the beginscreen and handels the actual game.
+     * 
      * @param args
      * @throws InterruptedException
      */
@@ -146,85 +98,25 @@ public class SurvivalDriver extends Driver {
             Logger.log("LineUnavailableException", 7, 2);
             e.printStackTrace();
         }
-        
-        int startTime = (int) System.currentTimeMillis();
-        
         Logger.log("Main Frame created", 9, 4);
         driver = new SurvivalDriver(name);
-        Player player = new Player(name , 350, false);
-        
+        Player player = new Player(name, 350, false);
         game = GameCreator.createSurvival(player);
         score = new Score();
-        
         game.addPlayer(player);
-        
         player = game.getPlayerList().get(0);
-        
         int centerConstant = (int) Math
                 .round(0.5 * (Settings.getScreenWidth() - Settings
                         .getLevelWidth()));
         Settings.setLeftMargin(centerConstant);
         GameScreen.setupScreen(game, score);
-        
         driver.startScreen();
-
         LogSettings.setActiveLog(true);
-
         while (true) {
             if (game.inProgress()) {
                 game.update();
-                
+
                 curLevel = game.getCurrentLevel();
-                
-                int currentTime = (int) System.currentTimeMillis();
-                if ((currentTime - startTime) > spawnTime) {
-                    spawnTime = spawnTime - 1000;
-                    
-                    score.addScore(100);
-                    
-                    if (allBubbles == 2) {
-                        spawnTime = 20000;
-                    }
-                    
-                    if (allBubbles == 3) {
-                        spawnTime = 30000;
-                    }
-                    
-                    startTime = (int) System.currentTimeMillis();
-                    
-                    for (int i = 0; i < allBubbles; i++) {
-                        randomPlacedBubble(bubbleNumber);
-                        switch (bubbleNumber) {
-                        case 1:
-                            curLevel.addBubble(bubble1);
-                            break;
-                        case 2:
-                            curLevel.addBubble(bubble2);
-                            break;
-                        case 3:
-                            curLevel.addBubble(bubble3);
-                            break;
-                        case 4:
-                            curLevel.addBubble(bubble4);
-                            break;
-                        case 5:
-                            curLevel.addBubble(bubble5);
-                            break;
-                        default:
-                            curLevel.addBubble(bubble1);
-                            break;
-                        }
-                    }
-                    
-                    if (bubbleNumber == 5) {
-                        bubbleNumber = 1;
-                        allBubbles++;
-                    }
-                    else {
-                        bubbleNumber++;
-                    }
-                    
-                }
 
                 for (int i = 0; i < curLevel.getBubbleList().size(); i++) {
                     Bubble bubble = curLevel.getBubbleList().get(i);
@@ -233,14 +125,14 @@ public class SurvivalDriver extends Driver {
 
                 for (int i = 0; i < curLevel.getPowerupList().size(); i++) {
                     curLevel.getPowerupList().get(i).move();
-                    curLevel.checkPowerupCollision();
+                    curLevel.handlePowerupCollision();
                 }
 
                 if (curLevel.hasRope()) {
                     curLevel.getRope().move();
                 }
 
-                curLevel.checkCollisionRope();
+                curLevel.handleCollisionRope();
 
                 if (curLevel.checkCollisionPlayer()) {
                     game.getPlayerList().get(0).removeAllPowerUps();
@@ -278,7 +170,9 @@ public class SurvivalDriver extends Driver {
                 Player player1 = game.getPlayerList().get(0);
                 player1.move();
 
-                checkGameLost();
+                if (checkGameLost()) {
+                    gameLost();
+                }
             }
 
             // 120 FPS
@@ -290,26 +184,6 @@ public class SurvivalDriver extends Driver {
             }
         }
 
-    }
-    
-    
-    /**
-     * Initialises the game..
-     */
-    public void initGame() {
-        try {
-            gameScreen = new GameScreen();
-        } catch (UnsupportedAudioFileException e) {
-            Logger.log("UnsupportedAudioFileException", 7, 2);
-            e.printStackTrace();
-        } catch (IOException e) {
-            Logger.log("IOException", 7, 2);
-            e.printStackTrace();
-        } catch (LineUnavailableException e) {
-            Logger.log("LineUnavailableException", 7, 2);
-            e.printStackTrace();
-        }
-        Logger.log("Main Frame created", 9, 4);
     }
 
     /**
@@ -328,5 +202,36 @@ public class SurvivalDriver extends Driver {
         Settings.setLeftMargin(centerConstant);
         GameScreen.setupScreen(game, score);
     }
-
+    /**
+     * initialize game.
+     */
+    public void initGame() {
+        try {
+            gameScreen = new GameScreen();
+        } catch (UnsupportedAudioFileException e) {
+            Logger.log("UnsupportedAudioFileException", 7, 2);
+            e.printStackTrace();
+        } catch (IOException e) {
+            Logger.log("IOException", 7, 2);
+            e.printStackTrace();
+        } catch (LineUnavailableException e) {
+            Logger.log("LineUnavailableException", 7, 2);
+            e.printStackTrace();
+        }
+        Logger.log("Main Frame created", 9, 4);
+    }
+    
+    public void startScreen() {
+        try {
+            new StartScreen(driver);
+        } catch (UnsupportedAudioFileException e1) {
+            e1.printStackTrace();
+        } catch (IOException e1) {
+            e1.printStackTrace();
+        } catch (LineUnavailableException e1) {
+            e1.printStackTrace();
+        }
+        Logger.log("Start screen created", 9, 4);
+    }
+    
 }
